Language classify [(#1095)](https://github.com/GoogleCloudPlatform/python-docs-samples/issues/1095)
* add classify text samples and tests

* use longer text

* move entity sentiment to v1

* flake

* year when first written

* year first written
