+++ samples/snippets/cloud-client/v1/snippets_test.py
@@ -1,4 +1,5 @@
-# Copyright 2016 Google, Inc.
+# -*- coding: utf-8 -*-
+# Copyright 2017 Google, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -56,3 +57,23 @@ def test_syntax_file(capsys):
     snippets.syntax_file(TEST_FILE_URL)
     out, _ = capsys.readouterr()
     assert 'NOUN: President' in out
+
+
+def test_sentiment_entities_text(capsys):
+    snippets.entity_sentiment_text(
+        'President Obama is speaking at the White House.')
+    out, _ = capsys.readouterr()
+    assert 'Content : White House' in out
+
+
+def test_sentiment_entities_file(capsys):
+    snippets.entity_sentiment_file(TEST_FILE_URL)
+    out, _ = capsys.readouterr()
+    assert 'Content : White House' in out
+
+
+def test_sentiment_entities_utf(capsys):
+    snippets.entity_sentiment_text(
+        'foo→bar')
+    out, _ = capsys.readouterr()
+    assert 'Begin Offset : 4' in out
