+++ samples/snippets/cloud-client/v1beta2/resources/android_text.txt
@@ -1,0 +1,1 @@
+Android is a mobile operating system developed by Google, based on the Linux kernel and designed primarily for touchscreen mobile devices such as smartphones and tablets.
